+++ packages/field/src/index.tsx
@@ -38,6 +38,7 @@ import FieldDigit from './components/Digit';
 import FieldSecond from './components/Second';
 import FieldRadio from './components/Radio';
 import FieldImage from './components/Image';
+import FieldCascader from './components/Cascader';
 import FieldColorPicker from './components/ColorPicker';
 import { noteOnce } from 'rc-util/lib/warning';
 
@@ -338,7 +339,9 @@ const defaultRenderText = (
   if (valueType === 'image') {
     return <FieldImage text={text as string} {...props} />;
   }
-
+  if (valueType === 'cascader') {
+    return <FieldCascader text={text as string} {...props} />;
+  }
   if (valueType === 'color') {
     return <FieldColorPicker text={text as string} {...props} />;
   }
@@ -356,7 +359,7 @@ export type ProFieldPropsType = {
 
 const ProField: React.ForwardRefRenderFunction<any, ProFieldPropsType> = (
   { text, valueType = 'text', onChange, renderFormItem, value, ...rest },
-  ref,
+  ref: any,
 ) => {
   const intl = useIntl();
   const context = useContext(ConfigContext);
@@ -377,9 +380,9 @@ const ProField: React.ForwardRefRenderFunction<any, ProFieldPropsType> = (
         text ?? fieldProps?.value ?? '',
         valueType || 'text',
         {
+          ref,
           ...rest,
           mode: rest.mode || 'read',
-          ref,
           renderFormItem: renderFormItem
             ? (...restProps) => {
                 const newDom = renderFormItem(...restProps);