+++ packages/form/src/demos/base.tsx
@@ -1,6 +1,7 @@
 import React, { useRef } from 'react';
 import { message } from 'antd';
 import type { ProFormInstance } from '@ant-design/pro-form';
+import { ProFormCascader } from '@ant-design/pro-form';
 import ProForm, {
   ProFormText,
   ProFormDateRangePicker,
@@ -41,7 +42,8 @@ export default () => {
         message.success('提交成功');
       }}
       formRef={formRef}
-      params={{}}
+      params={{ id: '100' }}
+      formKey="base-form-use-demo"
       request={async () => {
         await waitTime(100);
         return {
@@ -49,6 +51,7 @@ export default () => {
           useMode: 'chapter',
         };
       }}
+      autoFocusFirstInput
     >
       <ProForm.Group>
         <ProFormText
@@ -129,6 +132,46 @@ export default () => {
       <ProFormText width="sm" name="id" label="主合同编号" />
       <ProFormText name="project" width="md" disabled label="项目名称" initialValue="xxxx项目" />
       <ProFormText width="xs" name="mangerName" disabled label="商务经理" initialValue="启途" />
+      <ProFormCascader
+        width="md"
+        request={async () => [
+          {
+            value: 'zhejiang',
+            label: 'Zhejiang',
+            children: [
+              {
+                value: 'hangzhou',
+                label: 'Hangzhou',
+                children: [
+                  {
+                    value: 'xihu',
+                    label: 'West Lake',
+                  },
+                ],
+              },
+            ],
+          },
+          {
+            value: 'jiangsu',
+            label: 'Jiangsu',
+            children: [
+              {
+                value: 'nanjing',
+                label: 'Nanjing',
+                children: [
+                  {
+                    value: 'zhonghuamen',
+                    label: 'Zhong Hua Men',
+                  },
+                ],
+              },
+            ],
+          },
+        ]}
+        name="area"
+        label="区域"
+        initialValue={['zhejiang', 'hangzhou', 'xihu']}
+      />
     </ProForm>
   );
 };