+++ packages/layout/src/components/SiderMenu/BaseMenu.tsx
@@ -94,19 +94,21 @@ class MenuUtil {
 
   props: BaseMenuProps;
 
-  getNavMenuItems = (menusData: MenuDataItem[] = [], isChildren: boolean): React.ReactNode[] =>
-    menusData.map((item) => this.getSubMenuOrItem(item, isChildren)).filter((item) => item);
+  getNavMenuItems = (menusData: MenuDataItem[] = [], level: number): React.ReactNode[] =>
+    menusData.map((item) => this.getSubMenuOrItem(item, level)).filter((item) => item);
 
   /** Get SubMenu or Item */
-  getSubMenuOrItem = (item: MenuDataItem, isChildren: boolean): React.ReactNode => {
+  getSubMenuOrItem = (item: MenuDataItem, level: number): React.ReactNode => {
     if (Array.isArray(item.children) && item && item.children.length > 0) {
       const name = this.getIntlName(item);
       const { subMenuItemRender, prefixCls, menu, iconPrefixes, layout } = this.props;
       const isGroup = menu?.type === 'group' && layout !== 'top';
+      /** Menu 第一级可以有icon，或者 isGroup 时第二级别也要有 */
+      const hasIcon = level === 0 || (isGroup && level === 1);
       //  get defaultTitle by menuItemRender
       const defaultTitle = item.icon ? (
         <span className={`${prefixCls}-menu-item`} title={name}>
-          {!isChildren && getIcon(item.icon, iconPrefixes)}
+          {hasIcon && getIcon(item.icon, iconPrefixes)}
           <span className={`${prefixCls}-menu-item-title`}>{name}</span>
         </span>
       ) : (
@@ -116,18 +118,19 @@ class MenuUtil {
       );
 
       /** 如果是 Group 是不需要展示 icon 的 */
-      const subMenuTitle = isGroup ? (
-        <span className={`${prefixCls}-menu-item`} title={name}>
-          {name}
-        </span>
-      ) : (
-        defaultTitle
-      );
+      const subMenuTitle =
+        isGroup && level === 0 ? (
+          <span className={`${prefixCls}-menu-item`} title={name}>
+            {name}
+          </span>
+        ) : (
+          defaultTitle
+        );
       // subMenu only title render
       const title = subMenuItemRender
         ? subMenuItemRender({ ...item, isUrl: false }, defaultTitle)
         : subMenuTitle;
-      const MenuComponents: React.ElementType = isGroup && !isChildren ? ItemGroup : SubMenu;
+      const MenuComponents: React.ElementType = isGroup && level === 0 ? ItemGroup : SubMenu;
       return [
         <MenuComponents
           key={item.key || item.path}
@@ -140,7 +143,7 @@ class MenuUtil {
         >
           {this.getNavMenuItems(item.children, true)}
         </MenuComponents>,
-        isGroup && !isChildren && (
+        isGroup && level === 0 ? (
           <div
             key={`_${item.key || item.path}`}
             className={`${prefixCls}-menu-item-divider`}
@@ -154,13 +157,15 @@ class MenuUtil {
               }}
             />
           </div>
+        ) : (
+          false
         ),
       ];
     }
 
     return (
       <Menu.Item disabled={item.disabled} key={item.key || item.path} onClick={item.onTitleClick}>
-        {this.getMenuItemPath(item, isChildren)}
+        {this.getMenuItemPath(item, level)}
       </Menu.Item>
     );
   };
@@ -182,7 +187,7 @@ class MenuUtil {
    *
    * @memberof SiderMenu
    */
-  getMenuItemPath = (item: MenuDataItem, isChildren: boolean) => {
+  getMenuItemPath = (item: MenuDataItem, level: number) => {
     const itemPath = this.conversionPath(item.path || '/');
     const {
       location = { pathname: '/' },
@@ -193,8 +198,11 @@ class MenuUtil {
     } = this.props;
     // if local is true formatMessage all name。
     const name = this.getIntlName(item);
-    const { prefixCls } = this.props;
-    const icon = isChildren ? null : getIcon(item.icon, iconPrefixes);
+    const { prefixCls, menu } = this.props;
+    const isGroup = menu?.type === 'group';
+    /** Menu 第一级可以有icon，或者 isGroup 时第二级别也要有 */
+    const hasIcon = level === 0 || (isGroup && level === 1);
+    const icon = !hasIcon ? null : getIcon(item.icon, iconPrefixes);
     let defaultItem = (
       <span className={`${prefixCls}-menu-item`}>
         {icon}
@@ -416,7 +424,7 @@ const BaseMenu: React.FC<BaseMenuProps & PrivateSiderMenuProps> = (props) => {
       onOpenChange={setOpenKeys}
       {...props.menuProps}
     >
-      {menuUtils.getNavMenuItems(finallyData, false)}
+      {menuUtils.getNavMenuItems(finallyData, 0)}
     </Menu>
   );
 };