+++ packages/provider/src/locale/zh_TW.tsx
@@ -59,4 +59,7 @@ export default {
     open: '打開',
     close: '關閉',
   },
+  loginForm: {
+    submitText: '登入',
+  },
 };