+++ packages/utils/src/components/FilterDropdown/index.tsx
@@ -8,7 +8,7 @@ export type FooterRender =
   | ((
       onConfirm?: (e?: React.MouseEvent) => void,
       onClear?: (e?: React.MouseEvent) => void,
-    ) => JSX.Element)
+    ) => JSX.Element | false)
   | false;
 
 export type DropdownProps = {